+++ dialect/sqlite/dialect/update.go
@@ -18,6 +18,8 @@ type UpdateQuery struct {
 	clause.From
 	clause.Where
 	clause.Returning
+	clause.Limit
+	clause.Offset
 	bob.Load
 	bob.EmbeddedHook
 	bob.ContextualModdable[*UpdateQuery]
@@ -77,6 +79,20 @@ func (u UpdateQuery) WriteSQL(ctx context.Context, w io.Writer, d bob.Dialect, s
 		return nil, err
 	}
 	args = append(args, retArgs...)
+
+	limitArgs, err := bob.ExpressIf(ctx, w, d, start+len(args), u.Limit,
+		u.Limit.Count != nil, "\n", "")
+	if err != nil {
+		return nil, err
+	}
+	args = append(args, limitArgs...)
+
+	offsetArgs, err := bob.ExpressIf(ctx, w, d, start+len(args), u.Offset,
+		u.Offset.Count != nil, "\n", "")
+	if err != nil {
+		return nil, err
+	}
+	args = append(args, offsetArgs...)
 
 	return args, nil
 }