+++ dialect/sqlite/dm/qm.go
@@ -18,10 +18,35 @@ func From(name any) dialect.FromChain[*dialect.DeleteQuery] {
 	return dialect.From[*dialect.DeleteQuery](name)
 }
 
+func IndexedBy(i string) bob.Mod[*dialect.DeleteQuery] {
+	return bob.ModFunc[*dialect.DeleteQuery](func(q *dialect.DeleteQuery) {
+		q.From.IndexedBy = &i
+	})
+}
+
+func NotIndexed() bob.Mod[*dialect.DeleteQuery] {
+	return bob.ModFunc[*dialect.DeleteQuery](func(q *dialect.DeleteQuery) {
+		var s string
+		q.From.IndexedBy = &s
+	})
+}
+
 func Where(e bob.Expression) mods.Where[*dialect.DeleteQuery] {
 	return mods.Where[*dialect.DeleteQuery]{E: e}
 }
 
 func Returning(clauses ...any) bob.Mod[*dialect.DeleteQuery] {
 	return mods.Returning[*dialect.DeleteQuery](clauses)
+}
+
+func Limit(count any) bob.Mod[*dialect.DeleteQuery] {
+	return mods.Limit[*dialect.DeleteQuery]{
+		Count: count,
+	}
+}
+
+func Offset(count any) bob.Mod[*dialect.DeleteQuery] {
+	return mods.Offset[*dialect.DeleteQuery]{
+		Count: count,
+	}
 }